Simple implementation of a snake game
Snake considered by rectangles is moving in windows forms by using the
thick timer. In a random principle is decided where to put food for
snake( guid). The snake is growing bigger after eating a portion of food
(by adding a rectangle on every eaten food to the snake's tail). The
game is over when the snake went out of the border. For user interface i
have used Graphics class. User can easy manipulate the position of the
snake by keyboard arrows.